[skip changelog] Revert "Disable internal anchor checks by link checker tool (#1692)" (#2900)
This reverts commit f470f407a3afca82e30c6503b421ddde694c7b6d.

This configuration entry was previously required in order to work around a bug in the "markdown-link-check" tool that
resulted in false positives for links to anchors created by HTML anchor tags.

The bug in the "markdown-link-check" tool has since been fixed, and this project updated to use the version of the tool
with that fix. So this configuration entry is no longer required.

The now superfluous configuration entry is hereby removed in order to ensure more comprehensive link check coverage.