[skip changelog] Add missing pages to website navigation panel
The project includes a documentation website which is generated by the MkDocs static site generator. The site contains a
navigation panel which lists the available documentation pages.

Rather than being automatically generated from the contents of the `docs` folder, the contents of the navigation panel
have been explicitly defined via the MkDocs configuration file. This means that the project contributors must remember
to update the configuration file when development work results in the addition or removal of web pages.

This was not done when the `config get` and `debug check` commands were added. The web pages are generated as expected,
but the missing configuration file entries means that a visitor to the website would never know of their existence, and
could only access them by hacking the URL.

The missing entries are hereby added to the MkDocs configuration file.