Move information re: allowed characters for FQBN component to specifications (#2902)
Arduino CLI restricts which characters can be used by platform developers in their vendor, architecture, board, custom
board menu, and custom board option identifiers.

Previously, the information was only documented(?) in the "FAQ". This type of requirement must be present in the formal
specifications. A platform developer is unlikely to ever think to look for such information in the FAQ.

In addition, the inclusion of such information in the FAQ is not appropriate as this is intended for end users of the
Arduino CLI application and end users have no control over which characters were used by the platform developers in
their identifiers. So the presence of the information in the FAQ only harms readability by introducing unnecessary
complexity into a subject that users often struggle to understand.