[skip changelog] Correct documentation re: default for `build_cache.path` configuration key
The "Configuration" documentation page includes information on the default values of the configuration keys.

The default value of the build_cache.path configuration key was recently changed, but the documentation was not updated
accordingly.

The documentation is hereby updated to provide the correct information regarding the default value of the
build_cache.path configuration key.

Previously, the information was provided in the list of keys. That is a non-standard location for the information as
there is an established convention of instead providing this information in the separate section of the page dedicated
to documenting default paths. In addition. the location in the keys list became more problematic for the new default
since it is operating system-specific and thus the increased volume of documentation content would be awkward to fit
into the nested list of keys. For these reasons, the documentation content was moved to the standardized location under
the "Default directories" section of the page.