[skip changelog] Document Arduino CLI support for global platform.txt, and all recognized paths
Previously, the Platform Specification incorrectly stated that the "global platform.txt" feature is only supported by
Arduino IDE. In fact, it is also perfectly well supported when using Arduino CLI.

The cause of the misunderstanding is due to the documentation only mentioning that the file is recognized in the Arduino
IDE bundled packages path. Since that concept is specific to Arduino IDE 1.x (Arduino IDE 2.x did not exist at the time
this documentation was written, so the relevant IDE major version series was not specified), the assumption was made
that the feature was only available when using Arduino IDE 1.x. However, the system is actually general to all
packages folders, not exclusively the IDE bundled packages folder. The additional packages folders locations at which
global platform.txt files will be recognized by Arduino CLI and Arduino IDE 2.x, in addition to 1.x, are hereby
documented.