builder: RunRecipe: run either a set of recipes or a single recipe
Recipes must follow the 'prefix.NNN.suffix' pattern, where 'NNN' is a
number. When there is also a single 'prefix.suffix' property defined,
make sure to not include that in the list of recipes to run. However, if
no numbered recipes are found, use the single 'prefix.suffix' form if it
exists.

This allows to have both a single recipe and a set of numbered recipes
in the same build properties, for backwards compatibility.

Signed-off-by: Luca Burelli <l.burelli@arduino.cc>