linker: allow multi-step link recipe (`c.combine`) (#2954)
* builder: RunRecipe: run either a set of recipes or a single recipe

Recipes must follow the 'prefix.NNN.suffix' pattern, where 'NNN' is a
number. When there is also a single 'prefix.suffix' property defined,
make sure to not include that in the list of recipes to run. However, if
no numbered recipes are found, use the single 'prefix.suffix' form if it
exists.

This allows to have both a single recipe and a set of numbered recipes
in the same build properties, for backwards compatibility.

Signed-off-by: Luca Burelli <l.burelli@arduino.cc>

* linker: switch 'recipe.c.combine.pattern' to a recipe

Enable the use of recipes for the `recipe.c.combine.pattern` command.
Allows for more flexibility in the build process.

Signed-off-by: Luca Burelli <l.burelli@arduino.cc>

* Removed useless cloning of buildProperties

* Added unit test for recipe finder

---------

Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
Co-authored-by: Cristian Maglie <c.maglie@arduino.cc>