Fixed ObjFileIsUptodate checks in library discovery.
During the library detection phase, we do not have objects files yet,
the only thing that we should check is if the dependency file is
up-to-date.

In this commit we let GCC create a dep file during the library discovery
phase.

We must ensure that the build-path folder exists otherwise the compiler
will not be able to create the dep file.