Interpret local library paths relative to the sketch folder
Previously a local library path would be relative to the sketch folder if you
passed a folder path to compile, but if you passed a path to a sketch .ino file,
it would be considered relative to that. This change makes it so that relative
paths will consistently be interpreted relative to the folder whether you
specify a .ino file or not.