fix: build.options now have the full path to used platforms (#3042)
Previously the path to the folder containing the platforms was used,
(for exmaple, `/home/cmaglie/.arduino15/packages` instead of
`/home/cmaglie/.arduino15/packages/arduino/hardware/avr/1.8.6`).
This kept the build.options unchanged even if a platform was updated,
leading to a possible cache-reuse even if a full rebuild is required.